\name{DataInsideBox}
\alias{DataInsideBox}
\title{returns subset of data within bounding box }
\description{returns subset of data within bounding box }
\usage{DataInsideBox(PBSbox, data, xycoords = c("X", "Y"), verbose = 0)}
\arguments{
  \item{PBSbox}{rectangle in PBSmapping format, i.e. at least 4 rows!}
  \item{data}{data to subset}
  \item{xycoords}{spatial coordinates names}
  \item{verbose}{level of verbosity}
}



\author{Markus Loecher}




\examples{
#examples to come
print(1)
}
