\name{list.rules}
\alias{list.rules}
\title{compute rules for terminal nodes}
\description{extract rules for leaf nodes in both string and numeric form}
\usage{list.rules(x, i = NULL, verbose = 0)}
\arguments{
  \item{x}{object of class partykit}
  \item{i}{node number; if missing all nodes}
  \item{verbose}{level of verbosity}
}



\author{Markus Loecher}




\examples{
#examples to come
print(1)
}
