% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpartition.R
\name{plot_partition}
\alias{plot_partition}
\title{Plot the partitions of a tree model}
\usage{
plot_partition(xy, add = FALSE, col = NULL, lab = NULL, ...)
}
\arguments{
\item{xy}{object returned by \code{tree_partition}}

\item{add}{If true, add to existing plot, otherwise start a new plot.}

\item{col}{color}

\item{lab}{axis label}

\item{...}{Graphical parameters}
}
\value{
None
}
\description{
Plot the partitions of a tree involving one or two variables.
}
\examples{
\dontrun{
data(columbus, package="RgoogleMaps", envir = environment())
crimeTree <- tree(CRIME ~ X+Y, columbus)
xy=tree_partition(crimeTree,verbose=0)
plot_partition(xy,lab=TRUE)
}
}

