\name{PlotPartition}
\alias{PlotPartition}
\title{Plot the partitions of a tree model.}
\description{Plot the partitions of a tree involving one or two variables.}
\usage{PlotPartition(xy, add = FALSE, col = NULL, lab = NULL, ...)}
\arguments{
  \item{xy}{object returned by \code{TreePartition}}
  \item{add}{If true, add to existing plot, otherwise start a new plot.}
  \item{col}{color}
  \item{lab}{axis label}
  \item{\dots}{Graphical parameters}
}



\author{Markus Loecher}




\examples{
#examples to come
print(1)
}
