\name{PolyOverlap}
\alias{PolyOverlap}
\title{computes the relative overlap of two polygons}
\description{computes the relative overlap of two polygons}
\usage{PolyOverlap(polysA, polysB)}
\arguments{
  \item{polysA}{1st polygon}
  \item{polysB}{2nd polygon}
}

\value{relative polygon overlap measures}

\author{Markus Loecher}




\examples{
#examples to come
print(1)
}
