\name{Rect}
\alias{Rect}
\title{wrapper for base R rect function}

\usage{Rect(x, ...)}
\arguments{
  \item{x}{vector of length 4 with order xleft, ybottom, xright, ytop}
  \item{\dots}{further args to rect}
}



\author{Markus Loecher}




\examples{
#examples to come
print(1)
}
