\name{Rect2PBSpolys}
\alias{Rect2PBSpolys}
\title{simple utility function to convert object of rectangles into a PBSmapping type polygon}
\description{convert stacked rectangles into a labeled matrix}
\usage{Rect2PBSpolys(xy, xcols = c("X", "Y"))}
\arguments{
  \item{xy}{object returned by \code{TreePartition}}
  \item{xcols}{column names for x and y coordinates}
}

\value{PBSmapping type polygons}

\author{Markus Loecher}




\examples{
#examples to come
print(1)
}
