\name{Rules2BoundingBox}
\alias{Rules2BoundingBox}
\title{return bounding box of leaf node}

\usage{Rules2BoundingBox(ruleMatrix, bbox = list(lon = c(40, 60), lat = c(20, 
    30)), verbose = 0)}
\arguments{
  \item{ruleMatrix}{object returned by list.rules}
  \item{bbox}{replace Inf values with the corresponding bounding box limits}
  \item{verbose}{level of verbosity}
}



\author{Markus Loecher}




\examples{
#examples to come
print(1)
}
