\name{TreeHotspots-package}
\alias{TreeHotspots-package}
\alias{TreeHotspots}
\docType{package}
\title{Hotspot detection using classification trees}
}
\description{More about what it does (maybe more than one line)}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TreeHotspots}
\packageIndices{TreeHotspots}
}
\author{Markus Loecher}

Maintainer: \packageMaintainer{TreeHotspots}
}

\keyword{ package }


