% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rotate.R
\name{rotate}
\alias{rotate}
\title{rotate data by multiplying with a rotation matrix.}
\usage{
rotate(x, angle = 45, center = colMeans(x), inverse = FALSE)
}
\arguments{
\item{x}{data frame to be rotated}

\item{angle}{angle to rotate over}

\item{center}{center of rotation}

\item{inverse}{invert previously applied rotation}
}
\description{
rotate data by multiplying with a rotation matrix.
}
\examples{
# example to come
}

