\name{TreeHotspots-package}

\alias{TreeHotspots-package}

\alias{TreeHotspots}

\docType{package}

\title{Hotspot Detection using Classification Trees}


\description{Extracts leaves from trees that are candidates for hotspots.}



\details{

\tabular{ll}{Package: \tab TreeHotspots\cr

Type: \tab Package\cr

Title: \tab Hotspot detection using classification trees\cr

Version: \tab 1.0\cr

Date: \tab 2016-02-04\cr

Depends: \tab R (>= 2.10)\cr

Imports: \tab graphics, stats, tree, RgoogleMaps, PBSmapping, mvtnorm, partykit, rpart\cr

Suggests: \tab MASS\cr

Author: \tab Markus Loecher\cr

Maintainer: \tab Markus Loecher <markus.loecher@gmail.com>\cr

License: \tab GPL\cr

LazyLoad: \tab yes\cr}

}

\author{Markus Loecher}


\keyword{ package }







